% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arab}
\alias{arab}
\title{Arabidopsis RNA-seq Count Matrix}
\format{
A matrix with 26,222 rows and 6 columns.
}
\source{
Copied from the TCC package example data.
}
\description{
A count matrix used as the default seed population for
\code{\link[=sim_gene_counts]{sim_gene_counts()}}. Rows are genes and columns are samples.
}
\keyword{datasets}
