% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCountData.R
\name{as_DGEList}
\alias{as_DGEList}
\title{Convert to an edgeR DGEList}
\usage{
as_DGEList(x)
}
\arguments{
\item{x}{A \code{SeqCountData} object.}
}
\value{
An edgeR \code{DGEList}.
}
\description{
Convert to an edgeR DGEList
}
\examples{
if (requireNamespace("edgeR", quietly = TRUE)) {
    counts <- matrix(1:12, nrow = 3)
    colnames(counts) <- paste0("sample_", seq_len(ncol(counts)))
    rownames(counts) <- paste0("gene_", seq_len(nrow(counts)))
    exp_design <- data.frame(group = rep(c("A", "B"), each = 2))
    x <- newSeqCountData(
        counts,
        exp_design = exp_design,
        norm_factors = rep(1, ncol(counts))
    )
    d <- as_DGEList(x)
    d$samples$norm.factors
}

}
