% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{def_DEG}
\alias{def_DEG}
\title{Define Simulated DEGs}
\usage{
def_DEG(x, fc = 2)
}
\arguments{
\item{x}{A \linkS4class{SeqCountData} object whose \code{meta} slot contains a
\code{sim_params} entry of class \linkS4class{SimParams}.}

\item{fc}{Fold change cutoff. Defaults to 2.}
}
\value{
A logical vector with one value per gene.
}
\description{
Uses the fold change matrix stored in \code{x@meta$sim_params@fc} to define
simulated DEG status. A gene is marked \code{TRUE} when the largest between-group fold
change is at least \code{fc}.
}
\examples{
x <- sim_gene_counts()
head(def_DEG(x, fc = 4))

}
