% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_counts}
\alias{norm_counts}
\title{Get Normalized Counts}
\usage{
norm_counts(x)
}
\arguments{
\item{x}{A \code{SeqCountData} object.}
}
\value{
A normalized count matrix.
}
\description{
Returns normalized counts using effective library sizes.
}
\examples{
counts <- matrix(1:12, nrow = 3)
colnames(counts) <- paste0("sample_", seq_len(ncol(counts)))
rownames(counts) <- paste0("gene_", seq_len(nrow(counts)))
exp_design <- data.frame(group = rep(c("A", "B"), each = 2))
x <- newSeqCountData(
    counts,
    exp_design = exp_design,
    norm_factors = rep(1, ncol(counts))
)
norm_counts(x)

}
