% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqpatch-package.R
\docType{package}
\name{seqpatch-package}
\alias{seqpatch}
\alias{seqpatch-package}
\title{seqpatch}
\description{
Practical Patches for RNA-seq Count Data Analysis
}
\details{
For a comprehensive introduction to the package,
including a quick start and case studies,
see the package documentation site:
\url{https://bitdessin.github.io/seqpatch/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bitdessin/seqpatch}
  \item Report bugs at \url{https://github.com/bitdessin/seqpatch/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@bitdessin.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

}
\keyword{internal}
