% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCountData.R
\name{show,SeqCountData-method}
\alias{show,SeqCountData-method}
\title{Show a SeqCountData Object}
\usage{
\S4method{show}{SeqCountData}(object)
}
\arguments{
\item{object}{A \code{SeqCountData} object.}
}
\description{
Show a SeqCountData Object
}
\examples{
counts <- matrix(1:12, nrow = 3)
colnames(counts) <- paste0("sample_", seq_len(ncol(counts)))
rownames(counts) <- paste0("gene_", seq_len(nrow(counts)))
exp_design <- data.frame(group = rep(c("A", "B"), each = 2))
x <- newSeqCountData(counts, exp_design = exp_design)
show(x)

}
