% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimParams.R
\docType{class}
\name{SimParams-class}
\alias{SimParams-class}
\title{Class to Store Simulation Parameters}
\value{
An object of class \linkS4class{SimParams}.
}
\description{
An S4 class for storing the parameters used to generate synthetic RNA-seq
count data with \code{\link[=sim_gene_counts]{sim_gene_counts()}}.
}
\section{Slots}{

\describe{
\item{\code{n_genes}}{Number of genes.}

\item{\code{n_groups}}{Number of groups.}

\item{\code{n_replicates}}{Number of replicates per group.}

\item{\code{p_DEG}}{Proportion of genes assigned as DEGs for each group.}

\item{\code{fc_dist}}{Fold change distribution used for genes assigned as DEGs.}

\item{\code{fc_sd}}{Spread parameter used when \code{fc_dist} is not \code{"fixed"}.}

\item{\code{fc}}{Group-level fold change matrix. Rows are genes and columns are
groups.}

\item{\code{params_population}}{Data frame containing the seed-derived population of
mean, variance, dispersion, and distribution values.}

\item{\code{params}}{Data frame containing the sampled mean, variance, dispersion,
and distribution values used for each simulated gene.}
}}

\seealso{
\code{\link[=sim_gene_counts]{sim_gene_counts()}}
}
